% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawoRld-package.R
\docType{package}
\name{metawoRld-package}
\alias{metawoRld}
\alias{metawoRld-package}
\title{metawoRld: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anders Hagen Jarmund \email{anders.h.jarmund@ntnu.no} (\href{https://orcid.org/0000-0002-3923-1637}{ORCID})

}
\keyword{internal}
