% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{get_schema}
\alias{get_schema}
\title{Get the Schema from a metawoRld Project}
\usage{
get_schema(path = ".", schema = c("extraction", "assessment"))
}
\arguments{
\item{path}{Character string. The path to the root directory of the
metawoRld project. Defaults to the current working directory (\code{.}).}

\item{schema}{Character string. Either \code{extraction} (default) or \code{assessment}.}
}
\value{
A list representing the schema defined in the project's
\verb{_metawoRld.yml} file. This typically includes definitions for
\code{metadata_fields} and \code{data_fields}. Returns \code{NULL} invisibly and issues
a warning if the schema cannot be retrieved.
}
\description{
Reads the \verb{_metawoRld.yml} configuration file from a project
directory and returns the defined schema.
}
